Address serializer field consistency re: interchange methods
